controller update

+++ app/Services/BlogService.php
@@ -1,15 +1,16 @@
 <?php
+
 namespace App\Services;
 
 use App\Models\Blog;
 use Illuminate\Http\Exceptions\HttpResponseException;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Validator;
 use PDOException;
 
 class BlogService {
-
-    public function __construct(private Blog $blog){}
+    public function __construct(private Blog $blog) {}
 
     public function create(array $blog): Blog {
         try {
@@ -29,35 +30,53 @@ class BlogService {
         }
     }
 
-    public function getById(int $id): ?Blog {
-        try {
-            return $this->validateBlog($id);
-        } catch (PDOException $e) {
-            Log::error($e->getMessage());
-            throw $this->databaseException();
+    public function getById($id): ?Blog
+{
+    try {
+        $blog = $this->blog->find($id);
+        if (!$blog) {
+            throw $this->invalidException();
         }
+        return $blog;
+    } catch (PDOException $e) {
+        Log::error($e->getMessage());
+        throw $this->databaseException();
     }
+}
 
-    public function update(int $id, array $data): Blog {
-        try {
-            $blog = $this->validateBlog($id);
-            $blog->update($data);
-            return $blog;
-        } catch (PDOException $e) {
-            Log::error($e->getMessage());
-            throw $this->databaseException();
+public function update($id, array $data): Blog
+{
+    try {
+        $blog = $this->blog->find($id);
+        if (!$blog) {
+            throw $this->invalidException();
         }
+        $validatedData = Validator::make($data, [
+            'title' => 'string|max:255',
+            'description' => 'string',
+            'cover_image_url' => 'url',
+        ])->validate();
+        $blog->update($validatedData);
+        return $blog;
+    } catch (PDOException $e) {
+        Log::error($e->getMessage());
+        throw $this->databaseException();
     }
+}
 
-    public function delete(int $id): void {
-        try {
-            $blog = $this->validateBlog($id);
-            $blog->delete();
-        } catch (PDOException $e) {
-            Log::error($e->getMessage());
-            throw $this->databaseException();
+public function delete($id): void
+{
+    try {
+        $blog = $this->blog->find($id);
+        if (!$blog) {
+            throw $this->invalidException();
         }
+        $blog->delete();
+    } catch (PDOException $e) {
+        Log::error($e->getMessage());
+        throw $this->databaseException();
     }
+}
 
     private function databaseException(): HttpResponseException {
         return new HttpResponseException(
@@ -75,13 +94,5 @@ class BlogService {
                 'message' => 'Blog does not exist.'
             ], 404)
         );
-    }
-
-    private function validateBlog(int $id): Blog {
-        $blog = $this->blog->with('posts')->find($id);
-        if (!$blog) {
-            throw $this->invalidException();
-        }
-        return $blog;
     }
 }
